+++ noodles-gff/src/feature/record_buf.rs
@@ -6,6 +6,7 @@ mod convert;
 
 use std::io;
 
+use bstr::{BStr, ByteSlice};
 use noodles_core::Position;
 
 pub use self::{attributes::Attributes, builder::Builder};
@@ -174,16 +175,16 @@ impl Default for RecordBuf {
 }
 
 impl super::Record for RecordBuf {
-    fn reference_sequence_name(&self) -> &str {
-        self.reference_sequence_name()
+    fn reference_sequence_name(&self) -> &BStr {
+        self.reference_sequence_name().as_bytes().as_bstr()
     }
 
-    fn source(&self) -> &str {
-        self.source()
+    fn source(&self) -> &BStr {
+        self.source().as_bytes().as_bstr()
     }
 
-    fn ty(&self) -> &str {
-        self.ty()
+    fn ty(&self) -> &BStr {
+        self.ty().as_bytes().as_bstr()
     }
 
     fn feature_start(&self) -> io::Result<Position> {