+++ noodles-gff/src/io/writer/line/record/reference_sequence_name.rs
@@ -3,11 +3,12 @@ use std::{
     io::{self, Write},
 };
 
-use percent_encoding::{utf8_percent_encode, AsciiSet, NON_ALPHANUMERIC};
+use bstr::BStr;
+use percent_encoding::{AsciiSet, NON_ALPHANUMERIC};
 
 pub(super) fn write_reference_sequence_name<W>(
     writer: &mut W,
-    reference_sequence_name: &str,
+    reference_sequence_name: &BStr,
 ) -> io::Result<()>
 where
     W: Write,
@@ -18,7 +19,7 @@ where
 
 // § "Column 1: 'seqid'" (2020-08-18): "IDs may contain any characters, but must escape any
 // characters not in the set [a-zA-Z0-9.:^*$@!+_?-|]."
-fn percent_encode(s: &str) -> Cow<'_, str> {
+fn percent_encode(s: &BStr) -> Cow<'_, str> {
     const PERCENT_ENCODE_SET: &AsciiSet = &NON_ALPHANUMERIC
         .remove(b'.')
         .remove(b':')
@@ -33,7 +34,7 @@ fn percent_encode(s: &str) -> Cow<'_, str> {
         .remove(b'-')
         .remove(b'|');
 
-    utf8_percent_encode(s, PERCENT_ENCODE_SET).into()
+    percent_encoding::percent_encode(s, PERCENT_ENCODE_SET).into()
 }
 
 #[cfg(test)]
@@ -42,7 +43,7 @@ mod tests {
 
     #[test]
     fn test_write_reference_sequence_name() -> io::Result<()> {
-        fn t(buf: &mut Vec<u8>, reference_sequence_name: &str, expected: &[u8]) -> io::Result<()> {
+        fn t(buf: &mut Vec<u8>, reference_sequence_name: &BStr, expected: &[u8]) -> io::Result<()> {
             buf.clear();
             write_reference_sequence_name(buf, reference_sequence_name)?;
             assert_eq!(buf, expected);
@@ -51,10 +52,10 @@ mod tests {
 
         let mut buf = Vec::new();
 
-        t(&mut buf, "sq0", b"sq0")?;
-        t(&mut buf, "sq0.:^*$@!+_?-|", b"sq0.:^*$@!+_?-|")?;
-        t(&mut buf, "sq 0", b"sq%200")?;
-        t(&mut buf, ">sq0", b"%3Esq0")?;
+        t(&mut buf, BStr::new("sq0"), b"sq0")?;
+        t(&mut buf, BStr::new("sq0.:^*$@!+_?-|"), b"sq0.:^*$@!+_?-|")?;
+        t(&mut buf, BStr::new("sq 0"), b"sq%200")?;
+        t(&mut buf, BStr::new(">sq0"), b"%3Esq0")?;
 
         Ok(())
     }