+++ noodles-gtf/src/io/writer/line/record/reference_sequence_name.rs
@@ -1,13 +1,15 @@
 use std::io::{self, Write};
 
+use bstr::BStr;
+
 pub(super) fn write_reference_sequence_name<W>(
     writer: &mut W,
-    reference_sequence_name: &str,
+    reference_sequence_name: &BStr,
 ) -> io::Result<()>
 where
     W: Write,
 {
-    writer.write_all(reference_sequence_name.as_bytes())
+    writer.write_all(reference_sequence_name)
 }
 
 #[cfg(test)]
@@ -17,7 +19,7 @@ mod tests {
     #[test]
     fn test_write_reference_sequence_name() -> io::Result<()> {
         let mut buf = Vec::new();
-        write_reference_sequence_name(&mut buf, "sq0")?;
+        write_reference_sequence_name(&mut buf, BStr::new("sq0"))?;
         assert_eq!(buf, b"sq0");
         Ok(())
     }