+++ noodles-gtf/src/io/writer/line/record/source.rs
@@ -1,10 +1,12 @@
 use std::io::{self, Write};
 
-pub(super) fn write_source<W>(writer: &mut W, source: &str) -> io::Result<()>
+use bstr::BStr;
+
+pub(super) fn write_source<W>(writer: &mut W, source: &BStr) -> io::Result<()>
 where
     W: Write,
 {
-    writer.write_all(source.as_bytes())
+    writer.write_all(source)
 }
 
 #[cfg(test)]
@@ -14,7 +16,7 @@ mod tests {
     #[test]
     fn test_write_source() -> io::Result<()> {
         let mut buf = Vec::new();
-        write_source(&mut buf, "NDLS")?;
+        write_source(&mut buf, BStr::new("NDLS"))?;
         assert_eq!(buf, b"NDLS");
         Ok(())
     }